# SharpAI.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SharpAI.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/whisper-load": {
      "post": {
        "tags": [
          "Onnx"
        ],
        "parameters": [
          {
            "name": "useCudaDeviceId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": -1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhisperModelInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WhisperModelInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WhisperModelInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/whisper-run": {
      "post": {
        "tags": [
          "Onnx"
        ],
        "parameters": [
          {
            "name": "audioId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transcribe",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "useTimestamps",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "IsCancellationRequested",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "CanBeCanceled",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "WaitHandle.Handle",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/IntPtr"
            }
          },
          {
            "name": "WaitHandle.SafeWaitHandle.IsInvalid",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "WaitHandle.SafeWaitHandle.IsClosed",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/whisper-run-stream": {
      "post": {
        "tags": [
          "Onnx"
        ],
        "parameters": [
          {
            "name": "audioId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transcribe",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "useTimestamps",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "IntPtr": {
        "type": "object",
        "additionalProperties": false
      },
...
      "WhisperModelInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "rootPath": {
            "type": "string",
            "nullable": true
          },
          "modelName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "sizeInMb": {
            "type": "number",
            "format": "double"
          },
          "encoderPath": {
            "type": "string",
            "nullable": true
          },
          "decoderPath": {
            "type": "string",
            "nullable": true
          },
          "tokenizerPath": {
            "type": "string",
            "nullable": true
          },
          "preprocessorConfigPath": {
            "type": "string",
            "nullable": true
          },
          "configPath": {
            "type": "string",
            "nullable": true
          },
          "generationConfigPath": {
            "type": "string",
            "nullable": true
          },
          "vocabPath": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "mergesPath": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}